requirements.txt
amqp==5.1.1
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
black==22.3.0
blinker==1.4
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.1
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
flake8==4.0.1
Flask==2.1.2
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
infinity==1.5
iniconfig==1.1.1
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytz==2022.1
redis==4.3.3
six==1.16.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.2
tomli==2.0.1
typing_extensions==4.2.0
validators==0.20.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
WTForms-Components==0.10.5
zipp==3.8.0
flask_mongoengine
flask_restx
bcrypt
python-dateutil
pyjwt